Fork JHttp
Needed to fork JHttp in order to pass authentication credentials for Transifex.  Thus far, only cURL is supported in the forked BDHttp package.  With further testing, to include implementation in the other drivers, these changes will be offered upstream.

As well, in the cURL driver, I changed how the getResponse method set the body and headers.  Instead of using the first set of headers, I've modified it so that it will use the last headers to report an accurate status.  This is necessary in the case of HTTP redirects, such as what happens with Transifex.

Lastly, the Transifex package has been updated to use the new BDHttp package.